@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -26,10 +26,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.3
   # if you are using latest flutter please update your flutter_svg package to ^0.18.*
-  flutter_svg: ^0.19.1 # help us to use SVG in our app
-  provider: ^4.3.1 # for State management
+  flutter_svg: ^0.22.0 # help us to use SVG in our app
+  provider: ^6.0.0 # for State management
 
 dev_dependencies:
   flutter_test:
