@@ -5,8 +5,8 @@ import '../size_config.dart';
 
 class CustomSurffixIcon extends StatelessWidget {
   const CustomSurffixIcon({
-    Key key,
-    @required this.svgIcon,
+    Key? key,
+    required this.svgIcon,
   }) : super(key: key);
 
   final String svgIcon;
