@@ -5,12 +5,12 @@ import '../size_config.dart';
 
 class DefaultButton extends StatelessWidget {
   const DefaultButton({
-    Key key,
+    Key? key,
     this.text,
     this.press,
   }) : super(key: key);
-  final String text;
-  final Function press;
+  final String? text;
+  final Function? press;
 
   @override
   Widget build(BuildContext context) {
@@ -24,9 +24,9 @@ class DefaultButton extends StatelessWidget {
           primary: Colors.white,
           backgroundColor: kPrimaryColor,
         ),
-        onPressed: press,
+        onPressed: press as void Function()?,
         child: Text(
-          text,
+          text!,
           style: TextStyle(
             fontSize: getProportionateScreenWidth(18),
             color: Colors.white,
