@@ -8,10 +8,10 @@ import '../size_config.dart';
 
 class ProductCard extends StatelessWidget {
   const ProductCard({
-    Key key,
+    Key? key,
     this.width = 140,
     this.aspectRetio = 1.02,
-    @required this.product,
+    required this.product,
   }) : super(key: key);
 
   final double width, aspectRetio;
