@@ -9,15 +9,15 @@ class Product {
   final bool isFavourite, isPopular;
 
   Product({
-    @required this.id,
-    @required this.images,
-    @required this.colors,
+    required this.id,
+    required this.images,
+    required this.colors,
     this.rating = 0.0,
     this.isFavourite = false,
     this.isPopular = false,
-    @required this.title,
-    @required this.price,
-    @required this.description,
+    required this.title,
+    required this.price,
+    required this.description,
   });
 }
 
