@@ -11,7 +11,7 @@ import 'product_images.dart';
 class Body extends StatelessWidget {
   final Product product;
 
-  const Body({Key key, @required this.product}) : super(key: key);
+  const Body({Key? key, required this.product}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
