@@ -7,13 +7,13 @@ import '../../../size_config.dart';
 
 class ProductDescription extends StatelessWidget {
   const ProductDescription({
-    Key key,
-    @required this.product,
+    Key? key,
+    required this.product,
     this.pressOnSeeMore,
   }) : super(key: key);
 
   final Product product;
-  final GestureTapCallback pressOnSeeMore;
+  final GestureTapCallback? pressOnSeeMore;
 
   @override
   Widget build(BuildContext context) {
