@@ -4,9 +4,9 @@ import '../../../size_config.dart';
 
 class TopRoundedContainer extends StatelessWidget {
   const TopRoundedContainer({
-    Key key,
-    @required this.color,
-    @required this.child,
+    Key? key,
+    required this.color,
+    required this.child,
   }) : super(key: key);
 
   final Color color;
