@@ -4,9 +4,9 @@ import '../../../size_config.dart';
 
 class SectionTitle extends StatelessWidget {
   const SectionTitle({
-    Key key,
-    @required this.title,
-    @required this.press,
+    Key? key,
+    required this.title,
+    required this.press,
   }) : super(key: key);
 
   final String title;
