@@ -5,11 +5,11 @@ import '../../../size_config.dart';
 
 class SplashContent extends StatelessWidget {
   const SplashContent({
-    Key key,
+    Key? key,
     this.text,
     this.image,
   }) : super(key: key);
-  final String text, image;
+  final String? text, image;
 
   @override
   Widget build(BuildContext context) {
@@ -25,12 +25,12 @@ class SplashContent extends StatelessWidget {
           ),
         ),
         Text(
-          text,
+          text!,
           textAlign: TextAlign.center,
         ),
         Spacer(flex: 2),
         Image.asset(
-          image,
+          image!,
           height: getProportionateScreenHeight(265),
           width: getProportionateScreenWidth(235),
         ),
