@@ -48,7 +48,7 @@ class Body extends StatelessWidget {
         TweenAnimationBuilder(
           tween: Tween(begin: 30.0, end: 0.0),
           duration: Duration(seconds: 30),
-          builder: (_, value, child) => Text(
+          builder: (_, dynamic value, child) => Text(
             "00:${value.toInt()}",
             style: TextStyle(color: kPrimaryColor),
           ),
